Split src/protein_quest/filters.py into modules per filter
