+++ src/protein_quest/filters/__init__.py
@@ -1,0 +1,1 @@
+"""Module for filtering structure files and their contents."""
